v1.0 beta Changelog : - Calculatrice polonaise fonctionnelle : calculs et clear etc.. - Eviter d'afficher les erreurs pour cela faire des try and catch partout ? à demander - nombre à virgule à implementer

+++ src/main/java/com/example/calculatrice_0/Pile.java
@@ -14,8 +14,8 @@ public class Pile{//héritage
         return pile;
     }
 
-    public int getLast(){//ne sert à r en vrai t'as raison
-        return (int) pile.get(pile.size()-1);
+    public double getLast(){//ne sert à r en vrai t'as raison
+        return (double) pile.get(pile.size()-1);
     }
 
     public void push(double nombre){
